+++ R/nlme.mmkin.R
@@ -24,7 +24,7 @@ get_deg_func <- function() {
 #' This functions sets up a nonlinear mixed effects model for an mmkin row
 #' object. An mmkin row object is essentially a list of mkinfit objects that
 #' have been obtained by fitting the same model to a list of datasets.
-#' 
+#'
 #' Note that the convergence of the nlme algorithms depends on the quality
 #' of the data. In degradation kinetics, we often only have few datasets
 #' (e.g. data for few soils) and complicated degradation models, which may